map nav-item data from json file
