+++ src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css"
 import "./components/common/styles.scss"
 import { Route, Routes } from "react-router-dom"
-import HomePage from "./components/pages/HomePage/HomePage"
+import HomePage from "./components/pages/HomePage/Index"
 import LoginScreen from "./components/pages/LoginScreen/LoginScreen"
 import SettingScreen from "./components/pages/SettingScreen/SettingScreen"
 import SignUpScreen from "./components/pages/SignUpScreen/SignUpScreen"
