+++ src/components/pages/HomePage/NavBar/NavSection/NavList/NavItems/NavItemSubMenu.jsx
@@ -1,0 +1,16 @@
+import React from "react"
+
+export default function NavItemSubMenu({ subMenu }) {
+	// console.log(subMenu);
+	return (
+		<ul className="nav__list-submenu show">
+			<li>
+				{subMenu.map((subI, sIndex) => {
+					console.log(subI)
+
+					return <a key={sIndex} href="#"> {subI.name} </a>
+				})}
+			</li>
+		</ul>
+	)
+}
